backend/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2
alembic
pydantic
python-dotenv
